refactor trip tests
